refactor: decompose Mind into focused submodules (RB-ARCH-007)
Split crates/core/src/mind.rs (747 lines) into:
- mind/mod.rs: struct, types, accessors
- mind/lifecycle.rs: open, recovery, permissions
- mind/locking.rs: cross-process file locking
- mind/read.rs: search, ask, timeline, get_context
- mind/write.rs: remember, save_session_summary
- mind/stats.rs: statistics computation and caching

Public API surface unchanged. All tests pass.
